refactor: split lib.rs into domain command modules
lib.rs 2,458→178 lines (-92.8%). Extracted commands/ (files, settings, diagnostics, usd, updater, alembic, integrations, bench, shot), state.rs, shared.rs.
