+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,9 @@
+pub(crate) mod alembic;
+pub(crate) mod bench;
+pub(crate) mod diagnostics;
+pub(crate) mod files;
+pub(crate) mod integrations;
+pub(crate) mod settings;
+pub(crate) mod shot;
+pub(crate) mod updater;
+pub(crate) mod usd;
